+++ extensions/wistia/src/my-medias.tsx
@@ -1,18 +1,20 @@
 import { showToast, popToRoot, List, Toast, ActionPanel, Action, Icon } from "@raycast/api";
-import { AccountInfo, WistiaApiError, WistiaMedia } from "./types";
+import { AccountInfo } from "./types";
 import { WistiaMediaListItem } from "./wistia-media-list-item";
 import { fetchAccountInfo, fetchMedias } from "./api";
 import { useCachedPromise, useCachedState } from "@raycast/utils";
 
 export default function Command() {
   const [accountInfo, setAccountInfo] = useCachedState<AccountInfo>("account-info");
-  
+
   const { isLoading, data: medias } = useCachedPromise(
     async () => {
       const [accountInfo, medias] = await Promise.all([fetchAccountInfo(), fetchMedias()]);
       setAccountInfo(accountInfo);
       return medias;
-    }, [], {
+    },
+    [],
+    {
       initialData: [],
       async onError(error) {
         if (error.message === "unauthorized_credentials") {
@@ -29,20 +31,25 @@ export default function Command() {
           });
         }
       },
-    }
-  )
+    },
+  );
 
   return (
-      <List
-        isLoading={isLoading}
-        searchBarPlaceholder="Filter by name or hashed id"
-      >
-        {!isLoading && !medias.length && <List.EmptyView icon="content-library.svg" title="0 Media" actions={<ActionPanel>
-          <Action.OpenInBrowser icon={Icon.Plus} title="Add Media" url={`${accountInfo?.url}/content/media`} />
-        </ActionPanel>} />}
-        {medias.map(
-          (media) => accountInfo && <WistiaMediaListItem key={media.id} media={media} accountInfo={accountInfo} />,
-        )}
-      </List>
+    <List isLoading={isLoading} searchBarPlaceholder="Filter by name or hashed id">
+      {!isLoading && !medias.length && (
+        <List.EmptyView
+          icon="content-library.svg"
+          title="0 Media"
+          actions={
+            <ActionPanel>
+              <Action.OpenInBrowser icon={Icon.Plus} title="Add Media" url={`${accountInfo?.url}/content/media`} />
+            </ActionPanel>
+          }
+        />
+      )}
+      {medias.map(
+        (media) => accountInfo && <WistiaMediaListItem key={media.id} media={media} accountInfo={accountInfo} />,
+      )}
+    </List>
   );
 }