+++ extensions/wistia/CHANGELOG.md
@@ -7,5 +7,6 @@
 - Add EmptyViews
 - Tweak error handling
 - Show all loading indicators
+- Display accessories properly for all media types
 
 ## [Initial Version] - 2022-01-12